Revert "Refactor Normalize() in the command pkg"
This reverts commit 7eb92e7bea7c6eb740a550ebf83063fc73c00728.
