+++ internal/command/config.go
@@ -1,34 +1,35 @@
 package command
 
 import (
+	"fmt"
+	"os/exec"
+	"path/filepath"
 	"strings"
+
+	"google.golang.org/protobuf/proto"
 
 	runnerv2alpha1 "github.com/stateful/runme/internal/gen/proto/go/runme/runner/v2alpha1"
 )
 
+type ErrUnsupportedLanguage struct {
+	langID string
+}
+
+func (e ErrUnsupportedLanguage) Error() string {
+	return fmt.Sprintf("unsupported language %s", e.langID)
+}
+
+type ErrInterpretersNotFound struct {
+	interpreters []string
+}
+
+func (e ErrInterpretersNotFound) Error() string {
+	return fmt.Sprintf("unable to look up any of interpreters %q", e.interpreters)
+}
+
 // Config contains a serializable configuration for a command.
 // It's agnostic to the runtime or particular execution settings.
 type Config = runnerv2alpha1.ProgramConfig
-
-type configNormalizerCleanupFunc func() error
-
-type configNormalizer func(*Config) (*Config, configNormalizerCleanupFunc, error)
-
-func normalizeConfig(cfg *Config, normalizers ...configNormalizer) (_ *Config, cleanups []func() error, err error) {
-	for _, normalizer := range normalizers {
-		var cleanup func() error
-
-		cfg, cleanup, err = normalizer(cfg)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if cleanup != nil {
-			cleanups = append(cleanups, cleanup)
-		}
-	}
-	return cfg, cleanups, nil
-}
 
 // redactConfig returns a new Config instance and copies only fields considered safe.
 // Useful for logging.
@@ -43,7 +44,106 @@ func redactConfig(cfg *Config) *Config {
 	}
 }
 
-// TODO(adamb): this function is used for two quite different inputs: file extension and language ID.
+func normalizeConfig(cfg *Config, extra ...configNormalizer) (*Config, error) {
+	normalizers := []configNormalizer{
+		&pathNormalizer{},
+		&modeNormalizer{},
+	}
+
+	normalizers = append(normalizers, extra...)
+
+	for _, normalizer := range normalizers {
+		var err error
+
+		if cfg, err = normalizer.Normalize(cfg); err != nil {
+			return nil, err
+		}
+	}
+
+	return cfg, nil
+}
+
+type configNormalizer interface {
+	Normalize(*Config) (*Config, error)
+}
+
+type pathNormalizer struct{}
+
+func (n *pathNormalizer) Normalize(cfg *Config) (*Config, error) {
+	programPath, err := exec.LookPath(cfg.ProgramName)
+	if err == nil {
+		if programPath == cfg.ProgramName {
+			return cfg, nil
+		}
+
+		result := proto.Clone(cfg).(*Config)
+		result.ProgramName = programPath
+
+		return result, nil
+	}
+
+	interpreters := inferInterpreterFromLanguage(cfg.ProgramName)
+	if len(interpreters) == 0 {
+		return nil, &ErrUnsupportedLanguage{langID: cfg.ProgramName}
+	}
+
+	for _, interpreter := range interpreters {
+		program, args := parseInterpreter(interpreter)
+		if programPath, err := exec.LookPath(program); err == nil {
+			result := proto.Clone(cfg).(*Config)
+			result.ProgramName = programPath
+			result.Arguments = args
+			return result, nil
+		}
+	}
+
+	return nil, &ErrInterpretersNotFound{interpreters: interpreters}
+}
+
+type modeNormalizer struct{}
+
+func (n *modeNormalizer) Normalize(cfg *Config) (*Config, error) {
+	if cfg.Mode != runnerv2alpha1.CommandMode_COMMAND_MODE_UNSPECIFIED {
+		return cfg, nil
+	}
+
+	result := proto.Clone(cfg).(*Config)
+
+	if isShellLanguage(filepath.Base(result.ProgramName)) {
+		result.Mode = runnerv2alpha1.CommandMode_COMMAND_MODE_INLINE
+	} else {
+		result.Mode = runnerv2alpha1.CommandMode_COMMAND_MODE_FILE
+	}
+
+	return result, nil
+}
+
+func prepareScriptFromLines(programPath string, lines []string) string {
+	var buf strings.Builder
+
+	for _, cmd := range lines {
+		_, _ = buf.WriteString(cmd)
+		_, _ = buf.WriteRune('\n')
+	}
+
+	return buf.String()
+}
+
+func shellOptionsFromProgram(programPath string) (res string) {
+	base := filepath.Base(programPath)
+	shell := base[:len(base)-len(filepath.Ext(base))]
+
+	// TODO(mxs): powershell and DOS are missing
+	switch shell {
+	case "zsh", "ksh", "bash":
+		res += "set -e -o pipefail"
+	case "sh":
+		res += "set -e"
+	}
+
+	return
+}
+
 func isShellLanguage(languageID string) bool {
 	switch strings.ToLower(languageID) {
 	// shellscripts
@@ -68,4 +168,54 @@ func isShellLanguage(languageID string) bool {
 	default:
 		return false
 	}
+}
+
+// parseInterpreter handles cases when the interpreter is, for instance, "deno run".
+// Only the first word is a program name and the rest is arguments.
+func parseInterpreter(interpreter string) (program string, args []string) {
+	parts := strings.SplitN(interpreter, " ", 2)
+
+	if len(parts) > 0 {
+		program = parts[0]
+	}
+
+	if len(parts) > 1 {
+		args = strings.Split(parts[1], " ")
+	}
+
+	return
+}
+
+var interpreterByLanguageID = map[string][]string{
+	"js":              {"node"},
+	"javascript":      {"node"},
+	"jsx":             {"node"},
+	"javascriptreact": {"node"},
+
+	"ts":              {"ts-node", "deno run", "bun run"},
+	"typescript":      {"ts-node", "deno run", "bun run"},
+	"tsx":             {"ts-node", "deno run", "bun run"},
+	"typescriptreact": {"ts-node", "deno run", "bun run"},
+
+	"sh":          {"bash", "sh"},
+	"bash":        {"bash", "sh"},
+	"ksh":         {"ksh"},
+	"zsh":         {"zsh"},
+	"fish":        {"fish"},
+	"powershell":  {"powershell"},
+	"cmd":         {"cmd"},
+	"dos":         {"cmd"},
+	"shellscript": {"bash", "sh"},
+
+	"lua":    {"lua"},
+	"perl":   {"perl"},
+	"php":    {"php"},
+	"python": {"python3", "python"},
+	"py":     {"python3", "python"},
+	"ruby":   {"ruby"},
+	"rb":     {"ruby"},
+}
+
+func inferInterpreterFromLanguage(langID string) []string {
+	return interpreterByLanguageID[langID]
 }
